Add forcestart option to domsingester_init.sh
We want to be able to restart the ingester from ingest-object-mover but
obviously that is not possible when the init script now refuses to start
if ingest-object-mover is running.
As a workaround we add this undocumented option that bypasses the check
and allows ingest-object-mover to start the ingester as before.